Four more parameters spelled K where src/ said I
MSVC decorates `unsigned long` as K and `unsigned int` as I, and this tree
spells `uint32_t` as `unsigned int`. Same width, different decorated name, and
the name is what the linker pairs on - so four declarations disagreed with the
catalogue for a difference that changes no code:

    Wave::init(char *, uint32_t)   -> (char *, unsigned long)
    Wave::set_attrib(uint32_t)     -> (unsigned long)
    Net::get_player_name(uint32_t) -> (unsigned long)

None is in `src/OpenSMACX.def`, which is the filter that matters: two members
of this same family ARE exported, and changing one of those would drop a
pinned import from the staged image rather than fix anything.

Neither Wave row reaches BYTE_EXACT yet, and that is the point of doing it:
the C2511 that was hiding them is gone, and what the compiler now reports is
the defect underneath - a pointer conversion in `Wave::init`, and
`original_slot` undeclared in `Wave::set_attrib`. The second is a real
scaffolding gap, not a source defect: `original_slot` is a template, and the
preamble carries declarations rather than template definitions because a
template instantiation emits its own COMDAT and `byte_match.object_code`
refuses a unit holding two external symbols. Recorded here rather than
guessed at.

Also committed: the reusable report that found these, kept in the session
scratchpad rather than tools/ until it earns a test - it reads the live map,
joins each NO_COMPILE row to its decoded mangled name, and filters out the
rows whose fix would move an exported decoration.

+++ src/net_class.cpp
@@ -60,7 +60,7 @@ ORIGINAL: 0x00631A20
 Return Value: a pointer to the name, or nullptr
 Status: Complete
 */
-char *Net::get_player_name(uint32_t key) {
+char *Net::get_player_name(unsigned long key) {
     uint8_t *const bytes = reinterpret_cast<uint8_t *>(this);
     for (uint32_t index = 0; index < 0x10; ++index) {
         uint32_t candidate;
@@ -74,6 +74,6 @@ char *Net::get_player_name(uint32_t key) {
 }
 
 char *__fastcall net_get_player_name_redirect(
-        Net *self, void *, uint32_t key) {
+        Net *self, void *, unsigned long key) {
     return self->get_player_name(key);
 }
